split internal tests into two files
